Move terminal quick access into terminalContrib/
Fixes #230126
