+++ src/vs/workbench/contrib/terminal/terminal.all.ts
@@ -23,6 +23,7 @@ import '../terminalContrib/commandGuide/browser/terminal.commandGuide.contributi
 import '../terminalContrib/links/browser/terminal.links.contribution.js';
 import '../terminalContrib/zoom/browser/terminal.zoom.contribution.js';
 import '../terminalContrib/stickyScroll/browser/terminal.stickyScroll.contribution.js';
+import '../terminalContrib/quickAccess/browser/terminal.quickAccess.contribution.js';
 import '../terminalContrib/quickFix/browser/terminal.quickFix.contribution.js';
 import '../terminalContrib/typeAhead/browser/terminal.typeAhead.contribution.js';
 import '../terminalContrib/suggest/browser/terminal.suggest.contribution.js';
